requirements.txt
langchain-aws==0.1.0
boto3==1.38.4
pandas==2.1.4
python-dotenv==1.0.0
langchain==0.1.12
langchain-core==0.1.31
langchain-community==0.0.28